Updated post endpoints to include author information + verification (#190)
* Adding Update Comment Endpoint

* Associate user with new posts + verify user before updating or deleting

* Adding Post to Subthread collection + subthread to post collection

* Cleaning initial user check

* Fix non-existent user check

* Change error code for invalid user

* Fixing Test Cases

+++ server/src/routes/posts.js
@@ -2,6 +2,7 @@ const express = require('express');
 const { firebaseAuthMiddleware } = require('../middleware/firebaseAuth');
 
 const router = express.Router();
+const User = require('../db/models/users');
 const Post = require('../db/models/post');
 const Comment = require('../db/models/comments');
 const SubThread = require('../db/models/subThreads');
@@ -18,13 +19,26 @@ router.get('/', async (req, res) => {
 
 // Create one post
 router.post('/', firebaseAuthMiddleware, async (req, res) => {
-  const post = new Post({
-    title: req.body.title,
-    body: req.body.body,
-  });
-
   try {
+    let user = await User.findOne({ email: req.user.email });
+
+    if (!user) {
+      user = new User({
+        email: req.user.email,
+      });
+      await user.save();
+    }
+
+    const post = new Post({
+      title: req.body.title,
+      body: req.body.body,
+      author: user._id,
+      sub_thread: req.body.subthread,
+    });
+
     const newPost = await post.save();
+    await User.update({ _id: user._id }, { $push: { posts: newPost._id } });
+    await SubThread.update({ title: req.body.subthread }, { $push: { posts: newPost._id } });
     res.status(201).json(newPost);
   } catch (err) {
     res.status(400).json({ message: err.message });
@@ -35,11 +49,17 @@ router.post('/', firebaseAuthMiddleware, async (req, res) => {
 // eslint-disable-next-line no-unused-vars
 router.put('/:id', firebaseAuthMiddleware, async (req, res) => {
   try {
-    if (req.body.title != null && req.body.body != null) {
-      await Post.update({ _id: req.params.id }, { title: req.body.title, body: req.body.body });
-      res.status(200).send();
+    const currentPost = await Post.findById(req.params.id);
+    const user = await User.findById(currentPost.author);
+    if (user.email === req.user.email) {
+      if (req.body.title != null && req.body.body != null) {
+        await Post.update({ _id: req.params.id }, { title: req.body.title, body: req.body.body });
+        res.status(200).send();
+      } else {
+        res.status(400).json({ message: 'Please include a title and body' });
+      }
     } else {
-      res.status(400).json({ message: 'Please include a title and body' });
+      res.status(403).json({ message: 'You must be the author to update this post' });
     }
   } catch (err) {
     res.status(404).json({ message: err.message });
@@ -81,12 +101,23 @@ router.get('/:id', async (req, res) => {
 // eslint-disable-next-line no-unused-vars
 router.delete('/:id', firebaseAuthMiddleware, async (req, res) => {
   try {
-    await Post.findOneAndDelete({
-      _id: req.params.id,
-    });
-    res.status(200).send();
+    const currentPost = await Post.findById(req.params.id);
+    const user = await User.findById(currentPost.author);
+
+    if (user.email === req.user.email) {
+      try {
+        await Post.findOneAndDelete({
+          _id: req.params.id,
+        });
+        res.status(200).send();
+      } catch (err) {
+        res.status(404).json({ message: err.message });
+      }
+    } else {
+      res.status(403).json({ message: 'You must be the author to delete this post' });
+    }
   } catch (err) {
-    res.status(404).json({ message: err.message });
+    res.status(400).json({ message: err.message });
   }
 });
